Move cloud agnostic parts to engine package.
